+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to the "sfk-studio" extension will be documented in this file.
 
-## [Unreleased]
+## [1.2.0]
 
 ### Added
 - [TypeScript]: Compile with SFK transform on saving
@@ -37,6 +37,7 @@ All notable changes to the "sfk-studio" extension will be documented in this fil
 ## [1.0.0] - 2020-03-13
 - [GoToDefinition]: Initial feature
 
+[1.2.0]: https://github.com/phan2tom/vscode.sfk-studio/releases/tag/1.2.0
 [1.1.1]: https://github.com/phan2tom/vscode.sfk-studio/releases/tag/1.1.1
 [1.1.0]: https://github.com/phan2tom/vscode.sfk-studio/releases/tag/1.1.0
 [1.0.1]: https://github.com/phan2tom/vscode.sfk-studio/releases/tag/1.0.1